+++ thonny/backend.py
@@ -67,27 +67,27 @@ class BaseBackend(ABC):
             while True:
                 self._check_for_connection_error()
                 try:
-                    try:
-                        msg = self._fetch_next_incoming_message(timeout=0.01)
-                    except queue.Empty:
-                        self._perform_idle_tasks()
+                    msg = self._fetch_next_incoming_message(timeout=0.01)
+                except queue.Empty:
+                    self._perform_idle_tasks()
+                else:
+                    if isinstance(msg, InputSubmission):
+                        self._handle_user_input(msg)
+                    elif isinstance(msg, EOFCommand):
+                        self._handle_eof_command(msg)
                     else:
-                        if isinstance(msg, InputSubmission):
-                            self._handle_user_input(msg)
-                        elif isinstance(msg, EOFCommand):
-                            self._handle_eof_command(msg)
-                        else:
-                            self._current_command = msg
-                            self._handle_normal_command(msg)
-                except KeyboardInterrupt:
-                    self._send_output("KeyboardInterrupt", "stderr")  # CPython idle REPL does this
-                    self.send_message(ToplevelResponse())
-                except Exception:
-                    # Error in Thonny's code
-                    logger.exception("mainloop error")
-                    self._report_internal_exception("mainloop error")
+                        self._current_command = msg
+                        self._handle_normal_command(msg)
+        except KeyboardInterrupt:
+            self._send_output("KeyboardInterrupt", "stderr")  # CPython idle REPL does this
+            # TODO: is it safe to call it normal exit? And cause automatic reconnect?
+            sys.exit(0)
         except ConnectionError as e:
             self.handle_connection_error(e)
+        except Exception:
+            # Error in Thonny's code
+            logger.exception("mainloop error")
+            self._report_internal_exception("mainloop error")
 
         sys.exit(17)