+++ thonny/plugins/ev3/__init__.py
@@ -17,6 +17,25 @@ from thonny.plugins.micropython import (
 
 
 class EV3MicroPythonProxy(SshMicroPythonProxy):
+    def _get_launcher_with_args(self):
+        import thonny.plugins.ev3.ev3_back
+
+        args = {
+            "cwd": get_workbench().get_option(f"{self.backend_name}.cwd") or "",
+            "interpreter": self._target_executable,
+            "host": self._host,
+            "user": self._user,
+        }
+
+        args.update(self._get_time_args())
+        args.update(self._get_extra_launcher_args())
+
+        cmd = [
+            thonny.plugins.ev3.ev3_back.__file__,
+            repr(args),
+        ]
+        return cmd
+
     def _get_extra_launcher_args(self):
         return {"interpreter_launcher": ["brickrun", "-r", "--"]}
 
@@ -31,6 +50,7 @@ def load_plugin():
         EV3MicroPythonProxy,
         "MicroPython (EV3)",
         EV3MicroPythonConfigPage,
+        bare_metal=False,
         sort_key="23",
     )
     get_workbench().set_default("EV3MicroPython.executable", "pybricks-micropython")