+++ thonny/plugins/ev3/api_stubs/pybricks/hubs/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ["EV3Brick"]
+
+from .__stub.__ev3brick import EV3Brick