+++ thonny/plugins/ev3/api_stubs/pybricks/media/ev3dev/__init__.py
@@ -1,0 +1,6 @@
+__all__ = ["Font", "Image", "ImageFile", "SoundFile"]
+
+from .__stub.__imagefile import ImageFile
+from .__stub.__soundfile import SoundFile
+from .__stub.__font import Font
+from .__stub.__image import Image