+++ thonny/plugins/ev3/api_stubs/pybricks/parameters/__init__.py
@@ -1,0 +1,7 @@
+__all__ = ["Button", "Color", "Direction", "Port", "Stop"]
+
+from .__stub.__button import Button
+from .__stub.__color import Color
+from .__stub.__direction import Direction
+from .__stub.__port import Port
+from .__stub.__stop import Stop