+++ thonny/plugins/ev3/api_stubs/pybricks/tools/__init__.py
@@ -1,0 +1,5 @@
+__all__ = ["DataLog", "StopWatch", "wait"]
+
+from .__stub.__datalog import DataLog
+from .__stub.__stopwatch import StopWatch
+from .__stub.__tools import wait